removal and reorg of 'trackers' in preperation for publishing ofprovider public key to Coordinator.
