+++ nucypher/cli/painting/nodes.py
@@ -39,15 +39,6 @@ def paint_node_status(emitter, ursula, start_time):
 
     operator_address = 'Operator Address ...... {}'.format(ursula.operator_address)
     stats.extend([operator_address])
-
-    if ursula._availability_tracker:
-        if ursula._availability_tracker.running:
-            score = 'Availability Score .. {} ({} responders)'.format(ursula._availability_tracker.score, len(ursula._availability_tracker.responders))
-        else:
-            score = 'Availability Score .. Disabled'
-
-        stats.append(score)
-
     emitter.echo('\n' + '\n'.join(stats) + '\n')
