+++ nucypher/network/middleware.py
@@ -289,14 +289,6 @@ class RestMiddleware:
         )
         return response
 
-    def check_availability(self, initiator, responder):
-        response = self.client.post(node_or_sprout=responder,
-                                    data=bytes(initiator.metatada()),
-                                    path="check_availability",
-                                    timeout=6,  # Two round trips are expected
-                                    )
-        return response
-
     def ping(self, node):
         response = self.client.get(node_or_sprout=node, path="ping", timeout=2)
         return response
