+++ tests/conftest.py
@@ -7,16 +7,12 @@ from nucypher.blockchain.eth.actors import Operator
 from nucypher.blockchain.eth.networks import NetworksInventory
 from nucypher.crypto.powers import TransactingPower
 from nucypher.network.nodes import Learner
-from nucypher.network.trackers import AvailabilityTracker
 from nucypher.utilities.logging import GlobalLoggerSettings
 from tests.constants import MOCK_IP_ADDRESS, TESTERCHAIN_CHAIN_ID
 
 # Don't re-lock accounts in the background while making commitments
 LOCK_FUNCTION = TransactingPower.lock_account
 TransactingPower.lock_account = lambda *a, **k: True
-
-# Prevent halting the reactor via health checks during tests
-AvailabilityTracker._halt_reactor = lambda *a, **kw: True
 
 # Global test character cache
 global_mutable_where_everybody = defaultdict(list)
